changed get price to double

+++ src/main/java/com/pluralsight/Plaza/Pizzeria/toppings/PremiumMeatTopping.java
@@ -6,7 +6,7 @@ public class PremiumMeatTopping extends PremiumTopping{
     }
 
     @Override
-    public int getPrice(int size) {
+    public double getPrice(int size) {
         return 1;
     }
 }
